% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mean.R
\name{group_mean}
\alias{group_mean}
\title{Summarise mean value for each group}
\usage{
group_mean(data, var1, var2)
}
\arguments{
\item{data}{input dataset, including at least a column of grouping information and a column of measurements}

\item{var1}{the variable to group by, which can be condition/genotype/treament..}

\item{var2}{the variable to calculate means for}
}
\value{
a tibble with the group names, sample size of each group, and means for each group.
}
\description{
Like summarise(), group_mean()creates a new data frame. It's a combination of grouping by one variable and caculating measurements in each group. It's also useful in calculating an average for technical repeats, such as immunohistochemitry quantification (where you have multiple measurements for one sample).
}
\examples{
group_mean(data = PlantGrowth, var1 = group, var2 = weight)
}
